% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/catch_report.R
\name{catch_report}
\alias{catch_report}
\title{Catch report generator}
\usage{
catch_report(
  title,
  authors,
  stock,
  area,
  year,
  last_full_year,
  next_full_year,
  catch_data_date,
  output_dir = NULL
)
}
\arguments{
\item{title}{what do you want to call it?}

\item{authors}{who are the authors}

\item{stock}{a lowercase shorthand that matches with the data file https://docs.google.com/spreadsheets/d/1uHmCuY3GXfSBCbsP61nAQeATBfoslXeS3PQNLlioWIk/edit#gid=0}

\item{area}{lowercase typically "GOA" or "BSAI", though other "BSAI" areas can be called (e.g., Bogoslof) - must match the google sheet names}

\item{year}{current year}

\item{last_full_year}{last year that a full assessment was done}

\item{next_full_year}{next year that a full assessment is scheduled to occur}

\item{catch_data_date}{date that catch data are through, e.g., "November 5, 2023"}

\item{output_dir}{default output is the project, if you have a year folder the something like "2023" to id which folder to place the results in}
}
\description{
Catch report generator
}
\examples{
/dontrun{
catch_report(title = "1.B. Assessment of walleye pollock in the Bogoslof Island Region",
             authors = "James N. Ianelli, S. J. Barbeaux, Ivonne Ortiz, and D. McKelvey",
             stock = "pollock",
             area = "Bogoslof",
             year = 2023,
             last_full_year = 2021,
             next_full_year = 2024,
             catch_data_date = "November 5, 2023",
             output_dir = "2023")
}
}
