% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/read_specs.R
\name{read_specs}
\alias{read_specs}
\title{read in specs dataframe}
\usage{
read_specs(
  path,
  year,
  range = "A5:L91",
  col_names = c("species", "area", "0ofl", "0abc", "0tac", "0catch", "1ofl", "1abc",
    "1tac", "1catch", "2ofl", "2abc"),
  col_types = c("text", "text", rep("numeric", 10))
)
}
\arguments{
\item{path}{location of file}

\item{year}{current year}

\item{range}{the range of cells to include from the spreadsheet default:A5:L91}

\item{col_names}{specify names of all columns included}

\item{col_types}{specify column types of all columns included}
}
\description{
read in specs dataframe
}
\examples{
\dontrun{
read_specs('data/goa_specs2022.xlxs', col_names = c("species", "area", "0ofl", "0abc", "0tac", "0catch", "1ofl", "1abc", "1tac", "1catch", "2ofl", "2abc"), col_types = c("text", "text", rep("numeric", 10)))
}
}
