% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/spec_table.R
\name{spec_table}
\alias{spec_table}
\title{Create an Excel specs table}
\usage{
spec_table(year, month = "sep", area = "goa", catch_date = "11/9/2023")
}
\arguments{
\item{year}{current year}

\item{month}{sep or nov specs}

\item{area}{goa or bsai}

\item{catch_date}{date that catch dat is pulled e.g., ("9/10/2023")}
}
\description{
Create an Excel specs table
}
\examples{
/dontrun{
spec_table(year, month = "sep", area = "goa", catch_date="11/9/2023")
}

}
