% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{sum_tbl}
\alias{sum_tbl}
\title{age age plus group to biomass name}
\usage{
sum_tbl(spp, biomass, spec)
}
\arguments{
\item{spp}{species}

\item{biomass}{biomass data}

\item{spec}{specs data}
}
\description{
age age plus group to biomass name
}
