% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{sum_tbl_bs}
\alias{sum_tbl_bs}
\title{spec table bsai}
\usage{
sum_tbl_bs(spp, biomass, specs)
}
\arguments{
\item{spp}{species}

\item{biomass}{biomass data}

\item{specs}{specs data}
}
\description{
spec table bsai
}
